backend/requirements.txt
fastapi
uvicorn
supabase
openai
cohere
groq
tiktoken
python-dotenv
nomic[local]

